// src/interfaces/Player.ts
export interface Player {
    id:               number;
    nombre:           string;
    edad:             number;
    posicion:         string;
    dorsal:           number;
    goles:            number;
    asistencias:      number;
    faltas:           number;
    amarillas:        number;
    rojas:            number;
    minutos_jugados:  number;
    partidos_jugados: number;
    school:           number;
    equipo:           number;
}